Move remoteconfig.py from parametric to utils
